doddlebot v1.2 OVERHAUL Update
a overhaul to most of the back end has done. 

Help menus got an overhaul now with a new way to add colours and personal roles! Use d!help to have a look!

d!flip was added

A full SQL points system is in place 

cleverbot API was implemented

and a new role colour Aqua!